scrollToTop was preventing page usability
When the component initialy loaded, or updated, the scrollToTop function was called. The problem is that
if the page was already at the top the function prevented the user from scrolling down the page even if
the data was loaded. The function was changed to check to see if the page has been scrolled, and if so
by how much. Depending on the ammount the page has been scrolled, it adjusts the time accordingly.